requirements.txt
pynndescent==0.5.11
tqdm==4.64.1
scipy==1.11.4
numba==0.58.1
scikit-learn==1.3.2
matplotlib==3.6.3
pandas==2.1.3
numpy==1.24.2
umap-learn==0.5.5
